configure.ac: Improve compiler detection
Simplify setting clang/clang++ as the preferred compiler if CC is not set.

Import ax_compiler_vender.m4 from the autoconf-archive [1] and use the results
of AX_COMPILER_VENDOR to test whether clang++ is being used in order to avoid
false warnings when the binary is named something other than 'clang++'.

[1] https://www.gnu.org/software/autoconf-archive/ax_compiler_vendor.html